Basic motor rework.... again
