--- src/main/java/org/montclairrobotics/alloy/core/UniversalMotor.java
@@ -1,10 +1,0 @@
-package org.montclairrobotics.alloy.core;
-
-/**
- * Created by MHS Robotics on 2/25/2018.
- *
- * @author Garrett Burroughs
- * @since
- */
-public interface UniversalMotor extends TargetMotor, PoweredMotor {
-}
